# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMEN Signal Intelligence API",
    "description": "## Signal Intelligence Engine\n\nOMEN transforms market events into structured probability signals.\n\n### Authentication\n\nAll API endpoints require authentication via API key.\nInclude the `X-API-Key` header in all requests.\n\n**Public endpoints (no auth required):**\n- `/health/*` - Health checks\n- `/docs`, `/redoc` - API documentation\n- `/metrics` - Prometheus metrics\n\n### What OMEN provides\n- Probability assessment with confidence bounds\n- Signal validation and quality scoring\n- Geographic and temporal context\n- Full evidence chain and traceability\n\n### What OMEN does NOT provide\n- Impact assessment (severity, delay, cost)\n- Decision steering (temporal proximity, relevance rank, high-confidence flags)\n- Recommendations or advice\n- Risk quantification\n\nImpact assessment is the responsibility of downstream systems.",
    "version": "2.0.0",
    "x-logo": {
      "url": "https://omen.io/logo.png",
      "altText": "OMEN Logo"
    },
    "contact": {
      "name": "OMEN API Support",
      "url": "https://github.com/omen/omen",
      "email": "support@omen.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/api/v1/signals/stats": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "Get pipeline statistics",
        "description": "Get pipeline processing statistics for the signal-only path.\n\nReturns counts, pass rates, and latency metrics. No impact or recommendation data.\n\n**Example Response:**\n```json\n{\n    \"total_processed\": 1500,\n    \"total_passed\": 1250,\n    \"total_rejected\": 250,\n    \"pass_rate\": 0.833,\n    \"rejection_by_stage\": {\n        \"liquidity\": 150,\n        \"semantic\": 60,\n        \"geographic\": 40\n    },\n    \"latency_ms\": 45,\n    \"uptime_seconds\": 86400\n}\n```",
        "operationId": "get_signal_stats_api_v1_signals_stats_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineStatsResponse"
                },
                "example": {
                  "items": [
                    {
                      "signal_id": "sig-poly-12345",
                      "signal_type": "MARKET_MOVEMENT",
                      "title": "Polymarket: US Election probability shift",
                      "description": "Significant movement in election prediction markets",
                      "probability": 0.65,
                      "confidence": 0.88,
                      "keywords": [
                        "election",
                        "politics",
                        "prediction"
                      ],
                      "source": "polymarket",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "has_more": false,
                  "cursor": null
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/batch": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Create Signals Batch",
        "description": "Fetch events, run signal-only pipeline (validate → enrich → signal), return structured signals.\n\nThis is the Signal Intelligence contract: probability, confidence, context, evidence.\nNo impact calculations, no recommendations, no time-horizon or high-confidence flags.",
        "operationId": "create_signals_batch_api_v1_signals_batch_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "min_liquidity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1000,
              "title": "Min Liquidity"
            }
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.0,
              "title": "Min Confidence"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "List recent signals",
        "description": "List recent signals with pagination.\n\n**Query Parameters:**\n- `limit`: Max signals to return (default: 100, max: 1000)\n- `offset`: Pagination offset (default: 0)\n- `since`: Only return signals after this timestamp (ISO 8601)\n\n**Example Request:**\n```\nGET /api/v1/signals/?limit=10&since=2026-01-01T00:00:00Z\n```\n\n**Response includes:**\n- List of redacted signals\n- Pagination metadata (total, limit, offset)",
        "operationId": "list_signals_api_v1_signals__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "description": "Maximum number of signals to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of signals to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only return signals after this timestamp",
              "title": "Since"
            },
            "description": "Only return signals after this timestamp"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of signals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Signals Api V1 Signals  Get"
                },
                "example": {
                  "items": [
                    {
                      "signal_id": "sig-poly-12345",
                      "signal_type": "MARKET_MOVEMENT",
                      "title": "Polymarket: US Election probability shift",
                      "description": "Significant movement in election prediction markets",
                      "probability": 0.65,
                      "confidence": 0.88,
                      "keywords": [
                        "election",
                        "politics",
                        "prediction"
                      ],
                      "source": "polymarket",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "has_more": false,
                  "cursor": null
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/{signal_id}": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "Get signal by ID",
        "description": "Get a single signal by its unique identifier.\n\n**Detail Levels:**\n- `minimal`: Basic fields only (signal_id, title, probability, confidence_level)\n- `standard`: Default - includes evidence and geographic context\n- `full`: All fields including confidence breakdown\n\n**Example Request:**\n```\nGET /api/v1/signals/OMEN-A1B2C3D4E5F6?detail_level=full\n```",
        "operationId": "get_signal_api_v1_signals__signal_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          },
          {
            "name": "detail_level",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "minimal",
                "standard",
                "full"
              ],
              "type": "string",
              "description": "Level of detail to include in response",
              "default": "standard",
              "title": "Detail Level"
            },
            "description": "Level of detail to include in response"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signal details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Signal Api V1 Signals  Signal Id  Get"
                },
                "example": {
                  "signal_id": "OMEN-A1B2C3D4E5F6",
                  "source_event_id": "poly-12345",
                  "title": "Red Sea shipping disruption probability increases",
                  "probability": 0.72,
                  "probability_source": "polymarket",
                  "confidence_score": 0.85,
                  "confidence_level": "HIGH",
                  "confidence_factors": {
                    "liquidity": 0.9,
                    "geographic": 0.85,
                    "source_reliability": 0.8
                  },
                  "category": "GEOPOLITICAL",
                  "geographic": {
                    "regions": [
                      "Middle East",
                      "Red Sea"
                    ],
                    "chokepoints": [
                      "bab-el-mandeb",
                      "suez"
                    ]
                  },
                  "temporal": {
                    "event_horizon": "2026-06-30",
                    "resolution_date": "2026-06-30T00:00:00Z"
                  },
                  "evidence": [
                    {
                      "source": "Polymarket",
                      "source_type": "market",
                      "url": "https://polymarket.com/market/xyz"
                    }
                  ],
                  "trace_id": "a1b2c3d4e5f6g7h8",
                  "generated_at": "2026-02-01T12:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Signal not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "NOT_FOUND",
                  "message": "Signal 'OMEN-INVALID' not found",
                  "hint": "Verify the signal ID is correct and exists"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stats": {
      "get": {
        "tags": [
          "Statistics",
          "Statistics"
        ],
        "summary": "Get System Stats",
        "description": "Get current system statistics (signal-quality only).\n\nNo impact metrics. Values from real pipeline execution.",
        "operationId": "get_system_stats_api_v1_stats_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/multi-source/signals": {
      "get": {
        "tags": [
          "Multi-Source Intelligence"
        ],
        "summary": "Get Multi Source Signals",
        "description": "Get signals from multiple sources.\n\nSources:\n- polymarket: Prediction market signals\n- ais: Port congestion, chokepoint delays\n- weather: Storm alerts, sea conditions\n- freight: Rate spikes, capacity alerts\n\nSignals are combined and sorted by timestamp.",
        "operationId": "get_multi_source_signals_api_v1_multi_source_signals_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit_per_source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit Per Source"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated source names (e.g., 'ais,weather,freight')",
              "title": "Sources"
            },
            "description": "Comma-separated source names (e.g., 'ais,weather,freight')"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiSourceResponse"
                },
                "example": {
                  "items": [
                    {
                      "signal_id": "sig-poly-12345",
                      "signal_type": "MARKET_MOVEMENT",
                      "title": "Polymarket: US Election probability shift",
                      "description": "Significant movement in election prediction markets",
                      "probability": 0.65,
                      "confidence": 0.88,
                      "keywords": [
                        "election",
                        "politics",
                        "prediction"
                      ],
                      "source": "polymarket",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "has_more": false,
                  "cursor": null
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/multi-source/sources/{source_name}/signals": {
      "get": {
        "tags": [
          "Multi-Source Intelligence"
        ],
        "summary": "Get Single Source Signals",
        "description": "Get signals from a specific source.",
        "operationId": "get_single_source_signals_api_v1_multi_source_sources__source_name__signals_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "source_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token (optional, if enabled)"
      }
    }
  },
  "servers": [
    {
      "url": "https://api.omen.io",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.omen.io",
      "description": "Staging server"
...
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}